[56] UI menus should work
