+++ src/renderer/components/menu.js
@@ -1,7 +1,7 @@
-import game from '@/core/game'
-import filters from '@/core/filters'
+import game                from '@/core/game'
+import filters             from '@/core/filters'
 import ComponentMenuCursor from '@/components/menu_cursor'
-import electron from 'electron'
+import electron            from 'electron'
 
 const {ipcRenderer: ipc} = electron
 const {px} = filters
